Recursive exploration up: libc, thence gawk.
The errors from the bash build are so early they don't even clearly name the missing libs, but it's libc, of course.  So, we try to build that...

Building libc indicates that it requires gawk, so we recurse up to that...

Gawk here builds; we must now decide which things we're going to keep, how to package them... and, can we please get them built in such a way that it's relocatable?  We need a test formula for this, as well.  That's next.

You may notice that all three of these formulas so far have completely different paradigms for how the source gets brought in: gawk can just have it in the task dir; libc is different because its configure script (maturely, I feel, but apparently unusually for the ecosystem at large) refuses to run in the source dir; and bash had to be copied to another path to dance around permissions in the source package.  Se la vie, I guess?

Signed-off-by: Eric Myhre <hash@exultant.us>