Progress on glibc.  When the configure script says "don't run me from in the source dir", what that actually means is "you must run me from exactly one dir *inside* the source dir, because I have a bunch of '../' paths."
Thanks, whoever wrote that.  A couple more words there could have saved me S E V E R A L  H O U R S.

Now:

```
│ reppl eval >  /bin/sh: no: not found
```

I laughed out loud, tbh.

AMEND: REVISED STATEMENT.... this is not progress.  It's still broken.  Above theory completely wrong.

More context on that error:

```
│ reppl eval >  no gen-translit.pl < C-translit.h.in > C-translit.h.tmp
│ reppl eval >  /bin/sh: no: not found
│ reppl eval >  make[2]: *** [Makefile:76: C-translit.h] Error 127
│ reppl eval >  make[2]: *** Waiting for unfinished jobs....
│ reppl eval >  make[2]: Leaving directory '/task/locale'
│ reppl eval >  make[1]: *** [Makefile:215: locale/others] Error 2
│ reppl eval >  make[1]: Leaving directory '/task'
│ reppl eval >  make: *** [Makefile:9: all] Error 2
```

And *it's a race*.  The internationalization headers are still the dominant error.

I'm guessing "no" is also not, in fact, a command we should have, but rather that this is an error resulting from a configure script looking for "perl" or something and finding that "no" there is no perl :/

Long story short, we're still stuck on gettext here.

Signed-off-by: Eric Myhre <hash@exultant.us>